examples/medical_image/medical_pipeline.py
"""
Medical Image Analysis Pipeline
Based on COVID Chest X-ray Dataset
Repository: https://github.com/ieee8023/covid-chestxray-dataset
"""

import numpy as np
from typing import Dict, Any, Optional, List
import logging
from pathlib import Path
import requests
from PIL import Image
import io

logger = logging.getLogger(__name__)


class MedicalImagePipeline:
    """Medical image analysis pipeline for chest X-ray classification"""

    def __init__(self):
        self.model = None
        self.config = {
            "image_size": (224, 224),
            "classes": ["Normal", "COVID-19", "Pneumonia"],
            "data_url": "https://github.com/ieee8023/covid-chestxray-dataset",
        }

    def process_dicom_image(self, image_path: str) -> np.ndarray:
        """DICOM Image Processing"""
        try:
            # For demo, handle regular images
            # In production, use pydicom for DICOM files
            if image_path.endswith(".dcm"):
                # DICOM processing would go here
                # import pydicom
                # ds = pydicom.dcmread(image_path)
                # image = ds.pixel_array
                pass

            # Load and preprocess image
            img = Image.open(image_path)
            img = img.convert("RGB")
            img = img.resize(self.config["image_size"])

            # Normalize
            img_array = np.array(img) / 255.0
            img_array = np.expand_dims(img_array, axis=0)

            logger.info(f"✅ Image processed: {image_path}")
            return img_array
        except Exception as e:
            logger.error(f"Image processing failed: {e}")
            raise

    def augment_data(self, image: np.ndarray) -> List[np.ndarray]:
        """Data Augmentation"""
        augmented = [image]

        # Horizontal flip
        flipped = np.flip(image, axis=2)
        augmented.append(flipped)

        # Brightness adjustment
        bright = np.clip(image * 1.2, 0, 1)
        augmented.append(bright)

        # Dark adjustment
        dark = np.clip(image * 0.8, 0, 1)
        augmented.append(dark)

        logger.info(f"✅ Data augmentation: {len(augmented)} variants created")
        return augmented

    def train_cnn_model(self, images: List[np.ndarray], labels: List[int]):
        """CNN Model Training"""
        try:
            import tensorflow as tf
            from tensorflow import keras
            from tensorflow.keras import layers
        except ImportError:
            logger.warning("TensorFlow not available. Using placeholder model.")
            self.model = None
            return

        # Build CNN model
        model = keras.Sequential(
            [
                layers.Conv2D(
                    32,
                    (3, 3),
                    activation="relu",
                    input_shape=(*self.config["image_size"], 3),
                ),
                layers.MaxPooling2D((2, 2)),
                layers.Conv2D(64, (3, 3), activation="relu"),
                layers.MaxPooling2D((2, 2)),
                layers.Conv2D(64, (3, 3), activation="relu"),
                layers.Flatten(),
                layers.Dense(64, activation="relu"),
                layers.Dropout(0.5),
                layers.Dense(len(self.config["classes"]), activation="softmax"),
            ]
        )

        model.compile(
            optimizer="adam",
            loss="sparse_categorical_crossentropy",
            metrics=["accuracy"],
        )

        # Convert to arrays
        X = np.array(images)
        y = np.array(labels)

        # Train
        model.fit(X, y, epochs=10, batch_size=32, validation_split=0.2, verbose=0)

        self.model = model
        logger.info("✅ CNN model trained")
        return model

    def validate_model(
        self, images: List[np.ndarray], labels: List[int]
    ) -> Dict[str, Any]:
        """Model Validation"""
        if self.model is None:
            return {"error": "Model not trained"}

        try:
            import tensorflow as tf
        except ImportError:
            return {"error": "TensorFlow not available"}

        X = np.array(images)
        y = np.array(labels)

        # Evaluate
        loss, accuracy = self.model.evaluate(X, y, verbose=0)
        predictions = self.model.predict(X, verbose=0)

        return {
            "loss": float(loss),
            "accuracy": float(accuracy),
            "predictions": predictions.tolist(),
        }

    def predict(self, image_path: str) -> Dict[str, Any]:
        """Make prediction on medical image"""
        if self.model is None:
            return {
                "prediction": "Model not loaded",
                "confidence": 0.0,
                "error": "Please train the model first",
            }

        try:
            # Process image
            processed_image = self.process_dicom_image(image_path)

            # Predict
            prediction = self.model.predict(processed_image, verbose=0)[0]
            class_idx = np.argmax(prediction)
            confidence = float(prediction[class_idx])
            class_name = self.config["classes"][class_idx]

            return {
                "prediction": class_name,
                "confidence": confidence,
                "all_probabilities": {
                    self.config["classes"][i]: float(prediction[i])
                    for i in range(len(self.config["classes"]))
                },
            }
        except Exception as e:
            logger.error(f"Prediction error: {e}")
            return {"prediction": "Error", "confidence": 0.0, "error": str(e)}

    def clinical_deployment_checklist(self) -> Dict[str, Any]:
        """Clinical Deployment Checklist"""
        return {
            "model_validated": self.model is not None,
            "regulatory_compliance": {
                "hipaa_compliant": True,  # Placeholder
                "fda_approved": False,  # Placeholder
                "data_privacy": True,
            },
            "performance_metrics": {
                "accuracy_threshold": 0.95,
                "sensitivity_threshold": 0.90,
                "specificity_threshold": 0.90,
            },
            "deployment_status": "ready" if self.model else "not_ready",
        }
