pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-pipeline-deploy"
version = "1.0.0"
description = "Enterprise ML Pipeline Deployment Platform with 5 production-ready examples"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "ML Pipeline Team", email = "your-email@example.com"}
]
keywords = [
    "machine-learning",
    "ml-pipeline",
    "mlops",
    "deployment",
    "fastapi",
    "churn-prediction",
    "fraud-detection",
    "recommendation-system"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "scikit-learn==1.4.2",
    "pandas==2.2.2",
    "numpy==1.26.4",
    "fastapi==0.110.0",
    "uvicorn==0.28.0",
    "pydantic==2.6.4",
    "click==8.1.7",
    "pyyaml==6.0.1",
    "mlflow==2.11.1",
    "python-multipart==0.0.9",
    "rich==13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.2",
    "pytest-cov==4.1.0",
    "black==24.3.0",
    "flake8==7.0.0",
    "mypy==1.9.0",
]
full = [
    "torch==2.2.1",
    "transformers==4.38.2",
    "xgboost==2.0.3",
    "lightgbm==4.3.0",
    "pyarrow==15.0.0",
    "kedro==0.19.5",
    "prefect==2.18.1",
    "docker==6.1.3",
    "kubernetes==29.0.0",
    "boto3==1.34.59",
    "google-cloud-aiplatform==1.46.0",
    "azureml-core==1.55.0",
    "evidently==0.4.18",
    "prometheus-client==0.20.0",
    "grafana-api==1.0.3",
    "yfinance==0.2.37",
    "schedule==1.2.1",
    "kafka-python==2.0.2",
    "websockets==12.0",
    "tensorflow==2.16.1",
    "pydicom==2.4.4",
    "Pillow==10.2.0",
]

[project.urls]
Homepage = "https://github.com/Arvind-55555/ML-Pipeline-Deployment"
Documentation = "https://github.com/Arvind-55555/ML-Pipeline-Deployment#readme"
Repository = "https://github.com/Arvind-55555/ML-Pipeline-Deployment"
Issues = "https://github.com/Arvind-55555/ML-Pipeline-Deployment/issues"

[project.scripts]
ml-pipeline = "src.cli:main"
ml-pipeline-train = "run_local:train"
ml-pipeline-serve = "run_local:serve"
ml-pipeline-deploy = "run_unified_server:main"

[tool.setuptools]
packages = ["src", "examples", "src.core", "src.stages", "src.utils", "src.deployment"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.html", "*.js", "*.css"]

[tool.black]
line-length = 88
